move InputMode/VimMode types into input-mode.ts
